A minor issue for the replacing an order in a queue by admin(I should think for fixing this bug across all of my code :( so this is a better idea to use Restaurateur* instead of doing the exact work again by itself).

+++ src/Management/Restaurateur.cpp
@@ -222,6 +222,8 @@ bool Restaurateur::replaceOrderInQueue(string previousOrderID, string newOrderID
     RestaurantStorage rStorage;
     if(!storage.isValidOrder(newOrderID)) return false;
     if(!storage.isValidOrder(previousOrderID)) return false;
+    // if provious order is in queue and the new order, too:
+    if(rStorage.orderIsInQueue(restaurantID, previousOrderID) && rStorage.orderIsInQueue(restaurantID, newOrderID)) return false;
     return rStorage.deleteOrderFromRestaurant(restaurantID, previousOrderID) && rStorage.addOrderToRestaurant(restaurantID, newOrderID);
 }
 // statistics:
